refactor: Simplify JsonExtensions.
